Update README.md
Introduction:

If you are a data scientist and want to work with some real data, this article would help you create your own data set of job postings data. Here we will show how you can scrape the Internet to extract data and create your own real-life data set. The website that we are going to scrape is www.indeed.ca in order to extract job postings data.
We will get the job titles, company name, location, job posting URL, and more importantly, the job description.
For this purpose.

In this notebook we will use:

BeautifulSoup is a python library that lets you extract information from html and xml files. To see the documentation, pleae click here

Pandas for creating a data frame of the job postings data, and also for some data analysis

Matplotlib for the visualization. We will create some basic graphs in the end. If you are not already familiar with Matplotlib please click here

We will use the BeautifulSoup python library for the scraping part, and after we have the data set ready, we will do some data visualizations with Matplotlib.

Your questions and comments are welcome!

We will do this in five steps:

    Importing the libraries
    Creating a URL (a string) to brings us to the job posting site with the searching element we have defined
    Getting the number of jobs that are found as our search result. We need this to loop over each and every job posting's URL, and then get the information. We need to know how many jobs and how many pages we are going to scrape
    Extracting job postings information from the url and having them as a data frame
    Saving the data into a CSV file
    Doing some data cleansing
    Visualization of the distribution of the jobs by Canadian cities 

We have done some Natural Language Processing and some more data analysis using this data set. If you are interested, please take a look here